Initial functions to complete calculations
These functions each perform the different calculations for different types of loading checks that need to be done on a roof. I wanted the user to input the survey variables in one function which I would then have return these values when used as a variable in the other calculation functions. After setting this up I realised that this is the basic structure of a classes and sub classes so will adapt the code accordingly.